+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Docs: Added initial documentation drop
+- Feature: A script (`run-examples.sh`) to run and display the status (pass/fail) of each example
 
 ### Changed